+++ docs/docs/api/observers.mdx
@@ -1,8 +1,6 @@
----
-title: Observers
----
+## import { PubBadge } from '../../src/components/Shield';
 
-import { PubBadge } from '../../src/components/PubBadge';
+## title: Observers
 
 ## Observer Widget
