+++ docs/docs/core-concepts.mdx
@@ -1,11 +1,11 @@
----
-title: Core Concepts
-id: concepts
+## import { PubBadge } from '../src/components/Shield';
+
+title: Core Concepts id: concepts
+
 ---
 
 import mobxTriad from '../src/images/mobx-triad.png';
 import mobxIsometric from '../src/images/mobx-isometric.png';
-import { PubBadge } from '../src/components/PubBadge';
 
 <img
   src={mobxTriad}
