+++ docs/docs/examples/connectivity/index.mdx
@@ -1,9 +1,9 @@
----
-title: Connectivity
-slug: /examples/connectivity
+## import { PubBadge } from '../../../src/components/Shield';
+
+title: Connectivity slug: /examples/connectivity
+
 ---
 
-import { PubBadge } from '../../../src/components/PubBadge';
 import { Profile } from '../../../src/components/Testimonial';
 import connectivityGif from './connectivity.gif';
