+++ docs/docs/examples/dice/index.mdx
@@ -1,11 +1,11 @@
----
-title: Dice
-slug: /examples/dice
+## import { PubBadge } from '../../../src/components/Shield';
+
+title: Dice slug: /examples/dice
+
 ---
 
 import { Profile } from '../../../src/components/Testimonial';
 import diceJpg from './dice.gif';
-import { PubBadge } from '../../../src/components/PubBadge';
 
 <Profile
   author="Giri Jeedigunta"
