+++ docs/docs/examples/github/index.mdx
@@ -1,12 +1,10 @@
----
-title: GitHub Repos
-slug: /examples/github-repos
-description:
-  In this example, we will fetch the repositories for a github user. This will
-  also show the use of an ObservableFuture with an Observer.
----
+## import { PubBadge } from '../../../src/components/Shield';
 
-import { PubBadge } from '../../../src/components/PubBadge';
+title: GitHub Repos slug: /examples/github-repos description: In this example,
+we will fetch the repositories for a github user. This will also show the use of
+an ObservableFuture with an Observer.
+
+---
 
 In this example, we will use the <PubBadge name="github" /> package to fetch the
 repositories for a user. This will also show the use of an `ObservableFuture`
