+++ docs/docs/examples/todos/index.mdx
@@ -1,12 +1,10 @@
----
-title: Todos
-slug: /examples/todos
-description:
-  In this example, we will build the classic TodoMVC, which has become the
-  default way of showcasing the capabilities of a state-management library.
----
+## import { PubBadge } from '../../../src/components/Shield';
 
-import { PubBadge } from '../../../src/components/PubBadge';
+title: Todos slug: /examples/todos description: In this example, we will build
+the classic TodoMVC, which has become the default way of showcasing the
+capabilities of a state-management library.
+
+---
 
 In this example, we will build the classic [TodoMVC](http://todomvc.com/), which
 has become the default way of showcasing the capabilities of a state-management
