+++ docs/docs/guides/cheat-sheet.mdx
@@ -1,10 +1,9 @@
----
-slug: /guides/cheat-sheet
-title: MobX Cheat Sheet 🚀
-description: The MobX.dart cheat sheet
----
+## import { PubBadge } from '../../src/components/Shield';
 
-import { PubBadge } from '../../src/components/PubBadge';
+slug: /guides/cheat-sheet title: MobX Cheat Sheet 🚀 description: The MobX.dart
+cheat sheet
+
+---
 
 MobX has a pretty small API surface and tries to get out of your way as much as
 possible. Although minimal, there are few concepts you should be familiar with.
