+++ docs/docs/guides/json-serialization.mdx
@@ -1,9 +1,8 @@
----
-slug: /guides/json
-title: JSON Serialization of Stores
----
+## import { PubBadge } from '../../src/components/Shield';
 
-import { PubBadge } from '../../src/components/PubBadge';
+slug: /guides/json title: JSON Serialization of Stores
+
+---
 
 The <PubBadge name="json_serializable" /> package is a popular way to
 encode/decode between json representations of your models. It works by attaching
