+++ docs/docs/guides/organizing-stores.mdx
@@ -1,9 +1,9 @@
----
-slug: /guides/stores
-title: Organizing Stores
+## import { PubBadge } from '../../src/components/Shield';
+
+slug: /guides/stores title: Organizing Stores
+
 ---
 
-import { PubBadge } from '../../src/components/PubBadge';
 import storeHierarchy from './store-hierarchy.png';
 import dependencyOrder from './dependency-order.png';
