+++ docs/src/components/Sponsor.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { sponsors } from '../data/sponsors';
 
 export const Sponsor = ({ logo, url }) => {
   return (
@@ -8,23 +9,6 @@ export const Sponsor = ({ logo, url }) => {
   );
 };
 
-const sponsors = [
-  {
-    logo: require('../images/publicis-sapient-sponsor.png').default,
-    url: 'https://publicis.sapient.com',
-    active: false,
-  },
-  {
-    logo: require('../images/wunderdog-sponsor.png').default,
-    url: 'https://wunderdog.fi',
-    active: false,
-  },
-  {
-    logo: 'https://www.netlify.com/img/global/badges/netlify-color-bg.svg',
-    url: 'https://www.netlify.com',
-    active: true,
-  },
-];
 export const SponsorList = () => {
   return (
     <>
