+++ docs/src/data/sponsors.ts
@@ -1,0 +1,17 @@
+export const sponsors = [
+  {
+    logo: require('../images/publicis-sapient-sponsor.png').default,
+    url: 'https://publicis.sapient.com',
+    active: false,
+  },
+  {
+    logo: require('../images/wunderdog-sponsor.png').default,
+    url: 'https://wunderdog.fi',
+    active: false,
+  },
+  {
+    logo: 'https://www.netlify.com/img/global/badges/netlify-color-bg.svg',
+    url: 'https://www.netlify.com',
+    active: true,
+  },
+];
